Added Folder location check
Added the function:

To check if the page that the header got connected to is in the 'pages' folder or not. So it'll be ../js/header.js instead of js/header.js.

It has to exit out of the map first and then it's able to connect the js to the files in the pages map.